IRGen: simplify code generation for enum switch statements
[Cherry-picking test changes from this commit.]

This change uses the 'gather bits' functionality of enum payloads
to create a contiguous value to switch over. This allows us to
remove the code that currently attempts to build a switch statement
by comparing each element in the payload in turn.

The downside of this technique is that we may do more work up front
gathering bits and we may also need to compare larger values in some
situations. The upside is that we can remove a lot of complicated
code from IRGen. Also, we pass the responsibility for multi-way
branch generation to LLVM which can make use of a wider range of
switch lowering strategies than IRGen can sensibly support.

# Conflicts:
#	lib/IRGen/EnumPayload.cpp
#	test/IRGen/enum.sil