Sema: Synthesize accessors for computed local and global properties
If a stored property has a didSet/willSet, we currently synthesize
the getter and the setter inside the parser.  I want to get rid of
the parser's code path for doing that and consolidate all accessor
synthesis here in Sema.

Note that the parser did not synthesize a modify. This is now more
explicit in the code.

Also, we have to relax an assertion since addMemberToContextIfNeeded()
now gets called on declarations in local contexts.