[metadata] Change External::RelativeDirectPointer to be a real RelativeDirectPointer instead of an int32_t.
From talking with JoeG, this is due to historically RelativeDirectPointer not
having been ripped out of the runtime when this was written. This works now, so
it makes sense to change it. In terms of my goals, I want to use this to make it
easier to dump the names of protocols.