Treat 'some' types as having non-simple representations by default (#25620)
That is, if they appear inside a larger type, they may need to be
parenthesized. ASTPrinter is careful about this already, so this
doesn't actually change any existing behavior, but it makes the
default more conservative when no PrintOptions are in play.

(cherry picked from commit 4c26f95c01cc4b155061675869259433540b5845)