[AST] Increase size of type variable id field to 20 bits
Currently `id` field is only 12 bits long which is a lot smaller
than `graph index` which is 28 bits and creates problems with
large collections of literals since `id` is going to overflow
pretty quickly (4096 elements max). Let's change it so both
fields have the same 20 bits which is big enough.

Resolves: [SR-10914](https://bugs.swift.org/browse/SR-10914)
Resolves: rdar://problem/51616307