[build-script] Implement packaging in Python (only Linux)
Provide an alternative Python implementation of the packaging and
testing process for Linux. The implementation mostly translates
`build_and_test_installable_package` from the build-script-impl.

To make this change easier to check, only the Linux implementation is
provided and used. The macOS implementation can be provided later. If
necessary, a simple script to perform only the packaging is also
possible. Once the macOS implementation and the lipo step is performed
in Python, the build-script-impl code related to this parts can be
removed.

There are extra changes, specially around the arguments and implementing
the untar procedure.