[ModuleInterfaces] Clear dependencies before checking a different module
Sometimes, if one module we tried failed, we would leave the dependencies from the previous check and fail a different module which would have succeeded.

This manifested as the prebuilt-module-cache-forwarding test testing the wrong thing, because it retained a mod-time based dependency in the prebuilt cache, which should have been hash based, and used it.