[ClangImporter] Pass -skip-unused-modulemap-deps to clang
If there was a clang module in any of the search paths the swift module
loader was searching in, the clang dependency collector would include it
in our dependency tracker. This caused unfortunate build errors when
using .swiftinterface files from the SDK, because they would suddenly
depend on modulemaps in the benchmark suite.

rdar://51713745