[AutoDiff] Eliminate flow-sensitivity in '@differentiable' closure conversions. (#25688)
`@differentiable` function conversion from a non-`@differentiable` function is handled by inserting an `AutoDiffFunctionExpr` in CSApply when the contextual type is `@differentiable`. This is similar to C function poster conversion. One thing that we have not checked during these conversions is whether the original function expression is a literal closure expression or a reference to a `func` declaration. Lacking this check made `@differentiable` closure conversion flow-sensitive, which caused many semantic inconsistencies as described in [TF-577](https://bugs.swift.org/browse/TF-577).

This patch makes `@differentiable` closure conversion only valid when the from-expression is a literal closure expression or a reference to a `func` declaration. This eliminates all flow-sensitivity in `@differentiable` closure conversions.

In common cases where a user writes a higher-order function that differentiates the argument closure but forgets to make the parameter have a `@differentiable` or `@differentiable(linear)` function type, we produce a note and a fix-it to insert a `@differentiable` or `@differentiable(linear)` attribute to the parameter's type signature.