[TSan] Re-enable tsan-norace-deinit-run-time.swift
Second attempt for re-enabling this test. The reason for the previous
attempt failing was that I fixed the issue on the [swift-5.1-branch],
however, I was too eager to re-enable the test and did so before the fix
made it into the [stable] branch.

Fix:          78a6606f4 (Wed Jun 26 16:11:18 2019 -0700)
Merge commit: b3ec1d1c7 (Mon Jul  1 18:20:29 2019 -0700)

rdar://51804988
rdar://48680098