Driver: honour `-use-ld` on Windows static links
The librarian on Windows is a part of the linker.  Enabling `-use-ld=`
for driver for static linking on Windows enables the user to override
the linker.  This is particularly important for cross-linking Windows
from Linux where link.exe is not present as it is a part of the MSVC
toolset.