Driver: emit a diagnostic if `clang++` is not found
Rather than aborting due to an assertion failure, emit a diagnostic.
This is much safer and generally easier to understand why the command
failed.  It solves the problem of running swiftc from the build without
the path being set such that the clang++ driver is found by the swift
driver.