[typelowering] Workaround for rdar://52270675
The problem here is that in certain cases that I am investigating, given a
minimal resilience expansion, we are producing loadable, fragile types that have
subelements that are resilient =><=. The correct behavior is for this to be
address only.

So far, it seems we have only hit this problem around expanding types, so I am
putting in a small workaround while I investigate in order to unblock
Jordan/Doug.

rdar://52270675