SIL: Fix incorrect lowering of tuples containing trivial types which are resilient
If a type is trivial in its own module but resilient outside, then we
fail to propagate the 'is resilient' bit to the outer type's lowering.

As a result, if the first time the type is lowered with the Maximal
expansion, we don't re-compute the lowering when it is again lowered
with the Minimal expansion.

This can produce invalid SIL and trigger assertions.

Fixes <rdar://problem/52270675>.