IRGen: fix emitScatterBits when packed value is wider than target
When using an offset into the packed value it is reasonable for it
to be smaller than the target. For example, we might want to scatter
the high 8 bits of a 32 bit value into a 16 bit value. In this
scenario we mustn't just truncate the (32 bit) packed value to the
target (16 bit) type because we will lose the 8 bits we want to
scatter. We must therefore shift right first so that we can safely
truncate.

This fix allows us to avoid an explicit shift in interleaveSpareBits.