[diagnose-unreachable] Constant fold simple switch_enum_addr to eliminate more unreachable code.
This patch comes out of my reading some generic code using .none in transparent
functions to conditionally compile out code at -Onone. Sadly, before this the
dead code in question wouldn't be compiled out unless the protocol was
constrained to be a class protocol.

I added a test that validates that this conditional compilation property can be
relied on in -Onone code in both cases.