SIL: Fix SIL verifier's VerifyClassMethodVisitor
When checking a super method call, make sure we look at the vtable
for the right class.

Previously we would visit all the methods of the derived class,
together with any methods they override, but that is not quite
right -- it's perfectly legal to use 'super' to call a method that
you're not overriding too.

Fixes <https://bugs.swift.org/browse/SR-10260>.