[SE-0258] Remove incorrect assertion in SIL verifier for assign_by_wrapper.
The assertion was checking that the number of indirect result types
matched up with the number of indirect input types, which is... not
correct. The correct checks are later on, and the actual SIL itself is
correct. Fixes rdar://problem/52467175.

(cherry picked from commit ddb64a4af4013a0fad8e76e7c0b57c6c260851c3)