[RemoteMirror] Add support for QueryDataLayoutFunction to the interop header.
Remote Mirror added a new callback, QueryDataLayoutFunction, which provides a nicer and more extensible way to provide things like the target pointer size. Add support for that in the interop header. Search for and call the new creation function when it exists, and shim in support for older libraries that don't have it.

rdar://problem/52849572