Move SourceFile::LookupCache to header file
When using this in a different project, there are compiler errors that
result from std::default_delete being instantiated on a std::unique_ptr
on LookupCache, which is an incomplete type while in the header file.
This commit moves it to the header file so that it is a complete type.
There is one other similar error which will come in a future PR.